[UPDATE] 설정 스타일 변경
