+++ front/src/components/Footer/Footer.tsx
@@ -1,27 +1,24 @@
 import styled from 'styled-components';
+
 function Footer() {
   return (
     <FooterBlock>
-      <a href="https://github.com/LM-channel-team-project/team10">&#169;team10</a>
+      <ul>
+        <li>
+          &#169; 2021 LOGONAME.
+        </li>
+      </ul>
     </FooterBlock>
   );
 }
 
 const FooterBlock = styled.footer`
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  margin-bottom: 50px;
   font-size: 1.4rem;
-  a {
-    color: ${({ theme }) => theme['font-inactive']};
-    &:link,
-    &:visited {
-      text-decoration: none;
-    }
-    &:hover {
-      color: ${({ theme }) => theme['font']};
-      text-decoration: underline;
-    }
+  ul {
+    display: flex;
+    justify-content: center;
+    align-items: center;
   }
 `;
