+++ front/src/components/Settings/ToggleDarkmode/styles.ts
@@ -3,4 +3,9 @@ import styled from 'styled-components';
 export const Container = styled.div`
   display: flex;
   align-items: center;
+
+  span {
+    font-size: 1.6rem;
+    font-weight: 600;
+  }
 `;
