+++ front/src/pages/privacy.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function privacy() {
+	return (
+		<>
+		privacy
+		</>
+	)
+}
+
+export default privacy;
